Began segmentation
Began researching libraries to assist with tokenisation of user inputs in Chinese. Decided to employ the 'Jieba' library as it enabled functions to split inputs into base and further complicated phrases and terms. In the future, this would allow for more precise and a wider range of word recognition when comparing the input to tagging of pre-set responses within the program.