Completed first version of comparison scoring function
Completed the first version of the comparison scoring function. This function utilises the tokenise function to tokenise the input and existing follow up questions, before comparing them and scoring them as to which as the most consistently overlapping vocabulary. This gives a somewhat accurate representation of what follow-up questions will be most suitable for the user's input.