SQL Databases and SQLite
After learning how to setup SQL Databases and learning the basics of SQL, the main tables for underpinning questions, follow-up questions and categories have been established through the external DB Browser GUI for SQL. They are linked through primary and foreign keys so they are tagged appropriately through the categories. This database has been linked to the existing followupRelevancy function through SQLite, and now works properly to give a score on each of the follow-up questions in regard to the input stub.

Progress has also been made on the functions to add information to the follow-up question table in the database. This function takes user inputs as well as a determined questionID before adding it to the table.