Receiving follow-up terms based off categories
Added functionality to retrieve follow-up terms from the SQL database based off certain categories using SQLite. Also added the functionality to take the relevancy scores given to the follow-up passages and then match this to the original sentence, and then print the most accurate follow-up question depending on the user input and selected category of the input.