Beginning the mainline
I began a prototype command line interface to test the functionality of the program. It is able to select a random initial response from the category selected from the SQL database. It then can ask for an input from the user and give the best follow-up question in the database. A loop is necessary to enable conversation-like interaction but it is only a prototype and will need to be recreated when implemented into the GUI.