Signup and Login Functions
Creates functions for the signup and login for users. Just a placeholder to manage accounts so is unencrypted and stores the details in text in an SQL file. Can read and write into a file, scanning through existing users and checking if the accounts already exists and if the details are matching.