Moving GUI Libraries
Due to the inadequate and limiting capabilities and support for gooey pie, the library has been moved to Tkinter.
Tkinter allows for more functionality with in app windows and widgets that will make the design process for the GUI more efficient. There will also be less of a chance for unsupported issues and limits in the future of the program.