Finished Signup function
Attempted multiple different strategies for a secondary part to the signup function, where users input their account type as well as school and teacher. Made it so the type locks when a detail is selected so it cannot be broken before opening the next option box. It then through SQLite updates the user's details from a stored global userID variable set from the signup screen. It will be able to be updated and recalled through the main menu