Finished application
The chatbot functions correctly with capabilities to send email to the student's teacher, unless they are a teacher then it sends no email. The tabs switch flawlessly, passing the global userID. Teacher accounts have access to add questions to the database and tag them appropriately. There is no possible way to authenticate this really as the Chinese language is too extensive and there are too many other non-Chinese unicode characters to block, so it is based off trusting admin accounts.