Completed commenting bug-less code
Commented through all the code, removing redundant sections and checking for logic errors and inefficiencies. Could not find any logic errors and every place I could think an error could be I either fixed the authentication or added errors catches. There are some functions to do with verifying Chinese text that I cannot resolve as it is inefficient and almost impossible.

Also found a big where the signup function in the GUI was unable to check for spaces in the username and password, so I switched the authentication to a regex and added more comprehensive and user-friendly error messages.